+++ src/main/java/org/prgrms/voucherprgrms/voucher/model/VoucherDTO.java
@@ -1,19 +1,29 @@
 package org.prgrms.voucherprgrms.voucher.model;
 
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public class VoucherDTO {
-    private String voucherType;
-    private long value;
 
-    public String getVoucherType() {
-        return voucherType;
-    }
+    public final UUID voucherId;
 
-    public long getValue() {
-        return value;
-    }
+    public final String voucherType;
 
-    public VoucherDTO(String voucherType, long value) {
+    public final long value;
+
+    public final LocalDateTime createdAt;
+
+    public VoucherDTO(UUID voucherId, String voucherType, long value, LocalDateTime createdAt) {
+        this.voucherId = voucherId;
         this.voucherType = voucherType;
         this.value = value;
+        this.createdAt = createdAt;
+    }
+
+    public static VoucherDTO toVoucherDTO(Voucher voucher) {
+        return new VoucherDTO(voucher.getVoucherId(), voucher.getDTYPE(), voucher.getValue(), voucher.getCreatedAt());
     }
 }
